services/auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
PyJWT==2.8.0
mangum==0.17.0